chore: Update eslint configuration and page content
The eslint configuration has been updated to extend "next/core-web-vitals" and disable the "react/no-unescaped-entities" rule. Additionally, the content on the page has been modified to improve the description of the projects and add a credit section at the bottom.

Note: This is just a suggested commit message. Please modify it as needed to adhere to your repository's commit message conventions.